PIR: Remove TODO for pixels (#7209)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/488551667048375/task/1211877701264248?focus=true

### Description
Removes the TODO to add special handling for opt outs removed by user
when it comes to pixels. It's not needed as we mark that opt out as
deprecated and it's status is REMOVED_BY_USER meaning it won't be picked
up for pixel processing. Additionally, the related extracted profiles
are marked as deprecated, so they also won't be picked up.

### Steps to test this PR
QA is optional - you can run a new scan and wait until opt outs are
submitted, trigger the weekly pixel, then remove a record from dashboard
and trigger the weekly pixel again. The removed profile should not be
counted.

### UI changes
No UI changes