Fix: Wrong custom tab domain (#7249)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/488551667048375/task/1212224936885750?focus=true

### Description

This PR fixes a bug when the wrong domain is displayed if a user
navigates to a different domain URL within a custom tab.

See the recording:


https://github.com/user-attachments/assets/b11f036a-cbcb-4dab-a44e-30ec365d501e

Until now, the domain was only updated initially and when the page title
was updated, however, the URL used was not the one from the WebView but
the one in `Site`, which was not the latest one. That's why the current
domain shown was always one step behind the actual one.

I've refactored the custom tab view state updates so that the domain is
not set ad-hoc, but rather whenever an actual navigation change occurs.
Since the current logic used a domain that may be invalid for
determining the Duck Player state, I moved to the ViewState flow, as
well.

### Steps to test this PR

- [x] Open an app that opens a custom tab with a website that has links
to different domain sites
- [x] Tap on a link to navigate to a different domain
- [x] Notice the domain shown in the toolbar is correct