trust user-added CAs in Play store builds
updates network_security_config.xml in the Play store build to trust user-added CAs.  

This allows a user to use self-signed certs for https sites in a homelab or company intranet for example, provided the user has manually installed the root cert as a trusted CA via Android system settings.

Previously, if a user attempted to connect to an https site whose certificate was not issued by one of the standard CAs that ships with Android, then they would be presented with a "This site may be insecure" warning screen, even if the user had explicitly added the cert's root CA as a trusted one in the Android system settings.